refactor: JWT 및 쿠키 만료 시간 데이터 타입 변경 - JWT `accessExpiration`과 `refreshExpiration` 데이터 타입을 Integer에서 Long으로 변경 - 쿠키의 `maxAge` 데이터 타입을 Integer에서 Long으로 변경 - 관련 메서드 매개변수 데이터 타입 업데이트 - 코드 가독성 개선을 위한 소소한 포맷팅 수정

+++ src/main/java/com/example/moabackend/global/security/info/AuthenticationResponse.java
@@ -19,14 +19,9 @@ public class AuthenticationResponse {
             HttpServletResponse response,
             String domain,
             JwtDTO jwtDto,
-            Integer refreshExpiration
+            Long refreshExpiration
     ) throws IOException {
-//        CookieUtil.addCookie(
-//                response,
-//                domain,
-//                Constants.ACCESS_COOKIE_NAME,
-//                jwtDto.accessToken()
-//        );
+
         CookieUtil.addSecureCookie(
                 response,
                 domain,
@@ -61,7 +56,7 @@ public class AuthenticationResponse {
         response.setCharacterEncoding("UTF-8");
         response.setStatus(errorCode.getStatus().value());
 
-        Map<String, Object> body= new HashMap<>();
+        Map<String, Object> body = new HashMap<>();
         body.put("success", false);
         body.put("message", "요청이 실패했습니다.");
         body.put("data", null);
